+++ src/FAudioFX_reverb.c
@@ -1754,7 +1754,7 @@ uint32_t FAudioCreateReverb9WithCustomAllocatorEXT(
 void ReverbConvertI3DL2ToNative9(
 	const FAudioFXReverbI3DL2Parameters *pI3DL2,
 	FAudioFXReverbParameters9 *pNative,
-	uint8_t sevenDotOneReverb
+	int32_t sevenDotOneReverb
 ) {
 	float reflectionsDelay;
 	float reverbDelay;
